requirements.txt
openai>=1.60.0
boto3>=1.35.0
requests>=2.32.0
pyyaml>=6.0
